minimal updates

+++ src/common.h
@@ -1,3 +1,3 @@
 #include <time.h> 
 
-long int tdiff(struct timespec t1, struct timespec t2);
+double tdiff(struct timespec t1, struct timespec t2);
